refactor(core/types): move `Body` backwards-compatibility tests into package
